+++ src/main/java/ru/geekbrains/vlad/model/Category.java
@@ -3,7 +3,6 @@ package ru.geekbrains.vlad.model;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import ru.geekbrains.vlad.AbstractNamedEntity;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
